[ISPN-5526] Replication: The DELTA_WRITE flag should force a remote get during state transfer.
