+++ core/src/test/java/org/infinispan/atomic/DistAtomicMapStateTransferTest.java
@@ -1,0 +1,21 @@
+package org.infinispan.atomic;
+
+import org.infinispan.configuration.cache.CacheMode;
+import org.infinispan.transaction.TransactionMode;
+import org.testng.annotations.Test;
+
+/**
+ * Test modifications to an AtomicMap during state transfer are consistent across
+ * a distributed cluster.
+ *
+ * @author Ryan Emerson
+ * @author Dan Berindei
+ * @since 5.2
+ */
+@Test(groups = "functional", testName = "atomic.DistAtomicMapStateTransferTest")
+public class DistAtomicMapStateTransferTest extends BaseAtomicMapStateTransferTest {
+
+   public DistAtomicMapStateTransferTest() {
+      super(CacheMode.DIST_SYNC, TransactionMode.TRANSACTIONAL);
+   }
+}
